[SORT] move exceptoins to application layer
