+++ server/user-service/src/main/java/org/wid/userservice/application/service/oauth2/Oauth2Service.java
@@ -1,9 +1,9 @@
 package org.wid.userservice.application.service.oauth2;
 
 import org.springframework.web.reactive.function.client.ClientResponse;
-import org.wid.userservice.adapter.driving.rest.exception.BadRequestException;
 import org.wid.userservice.application.dto.oauth2.token.TokenResponseDto;
 import org.wid.userservice.application.dto.user.UserDto;
+import org.wid.userservice.application.exception.BadRequestException;
 
 import reactor.core.publisher.Mono;
